Add option to open as app window
This adds a question to the creation that asks whether to open in
Chrome’s “app” mode, which just removes the toolbars from the window.
This ends up with a simple window with a title and buttons, and no
tabs/toolbar icons/bookmarks bar, etc., making it look like a
standalone app.